+++ vitest.config.ts
@@ -5,5 +5,9 @@ export default defineConfig({
 		include: ["test/**/*.test.ts"],
 		maxWorkers: 2,
 		minWorkers: 2,
+		coverage: {
+			include: ["src/**"],
+			reporter: ["text", "html"],
+		},
 	},
 });